requirements.txt
pytest
selenium == 4.5.0
selene >= 2.0.0b17
allure-pytest